[Infra] Add Windows build support
This patch adds rudimentary windows build support. A complete
platform implementation is not included, so not all targets
are added to `gn_all` yet.

Bug: 236163082

+++ osp/impl/presentation/presentation_controller_unittest.cc
@@ -118,7 +118,7 @@ class ControllerTest : public ::testing::Test {
 
   void ExpectAvailabilityRequest(
       msgs::PresentationUrlAvailabilityRequest& request) {
-    ssize_t decode_result = -1;
+    int64_t decode_result = -1;
     msgs::Type msg_type;
     EXPECT_CALL(mock_callback_, OnStreamMessage(_, _, _, _, _, _))
         .WillOnce(Invoke([&request, &msg_type, &decode_result](
@@ -195,7 +195,7 @@ class ControllerTest : public ::testing::Test {
 
   void ExpectCloseEvent(MockMessageCallback* mock_callback,
                         Connection* connection) {
-    ssize_t decode_result = -1;
+    int64_t decode_result = -1;
     msgs::Type msg_type;
     msgs::PresentationConnectionCloseEvent event;
     EXPECT_CALL(*mock_callback, OnStreamMessage(_, _, _, _, _, _))
@@ -256,7 +256,7 @@ class ControllerTest : public ::testing::Test {
                              msgs::Type message_type, const uint8_t* buffer,
                              size_t buffer_size, Clock::time_point now) {
           msg_type = message_type;
-          ssize_t result = msgs::DecodePresentationStartRequest(
+          int64_t result = msgs::DecodePresentationStartRequest(
               buffer, buffer_size, request);
           return result;
         }));
@@ -404,7 +404,7 @@ TEST_F(ControllerTest, TerminatePresentationFromController) {
                            msgs::Type message_type, const uint8_t* buffer,
                            size_t buffer_size, Clock::time_point now) {
         msg_type = message_type;
-        ssize_t result = msgs::DecodePresentationTerminationRequest(
+        int64_t result = msgs::DecodePresentationTerminationRequest(
             buffer, buffer_size, termination_request);
         return result;
       }));
@@ -490,7 +490,7 @@ TEST_F(ControllerTest, Reconnect) {
       controller_->ReconnectConnection(std::move(connection),
                                        &reconnect_delegate);
   ASSERT_TRUE(reconnect_request);
-  ssize_t decode_result = -1;
+  int64_t decode_result = -1;
   msgs::Type msg_type;
   EXPECT_CALL(mock_callback, OnStreamMessage(_, _, _, _, _, _))
       .WillOnce(Invoke([&open_request, &msg_type, &decode_result](
